# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST entrypoint over the message classifier core. POST /api/v1/messages runs the same Classify → Dispatch as the web and the CLI and answers the outcome as JSON; the browse endpoints are read-only.",
        "title": "msg-classifier API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/contacts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "List contacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "all|phone|email|name",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "term matched against the name, the phone or the email",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Contact"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages": {
            "post": {
                "description": "Runs the same Classify → Dispatch core as the web and the CLI, and answers the outcome as JSON. `action` is the discriminant: contact_add, contact_found, contact_not_found, contact_duplicate, contact_no_data, note_add, note_found, note_not_found, note_no_data, note_duplicate, transaction_add, transaction_found, transaction_not_found, transaction_no_data, transaction_duplicate or none. On a duplicate the API only informs (DC-008): there is no confirmation path — re-posting the same message with `dup_action` set to `new` or `update` performs the confirmed save, but the web and CLI are the surfaces designed for that loop.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Classify a message",
                "parameters": [
                    {
                        "description": "Message to classify",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReceiveMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "List notes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "all|note|reminder|todo",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "term matched against the note content",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Note"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "all|compra|venda|pagamento|recebimento|transferencia",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "term matched against the party and the original message",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.MessageResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "classification": {
                    "$ref": "#/definitions/models.Classification"
                },
                "contact": {
                    "$ref": "#/definitions/models.Contact"
                },
                "message": {
                    "type": "string"
                },
                "missing": {
                    "type": "string"
                },
                "notes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Note"
                    }
                },
                "search_term": {
                    "type": "string"
                },
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SegmentScore"
                    }
                },
                "total": {
                    "type": "number"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
        "models.CategoryFinding": {
            "type": "object",
            "properties": {
                "choice": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                }
            }
        },
        "models.Classification": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.CategoryFinding"
                },
                "kind": {
                    "$ref": "#/definitions/models.KindFinding"
                }
            }
        },
        "models.Contact": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_norm": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.KindFinding": {
            "type": "object",
            "properties": {
                "choice": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                }
            }
        },
        "models.Note": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TodoItem"
                    }
                },
                "time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ReceiveMessageRequest": {
            "type": "object",
            "properties": {
                "dup_action": {
                    "description": "\"\" = check, \"new\" = insert anyway, \"update\" = merge into the duplicate",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.SegmentScore": {
            "type": "object",
            "properties": {
                "included": {
                    "type": "boolean"
                },
                "score": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.TodoItem": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "note_id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "content": {
                    "description": "the original message, kept verbatim",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "party": {
                    "description": "establishment or person involved",
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}